requirements.txt
# Embedding
FlagEmbedding>=1.2.0
sentence-transformers>=2.2.0
torch>=2.0.0

# Vector Database
qdrant-client>=1.7.0

# Hugging Face Inference API
huggingface-hub>=0.20.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
tqdm>=4.65.0

# Web App
streamlit>=1.30.0

# LLM
google-generativeai>=0.3.0

# Utilities
python-dotenv>=1.0.0
loguru>=0.7.0
